fix : Fixed issue where tags could not be cleared

+++ lib/page/collection/edit_collection_screen.dart
@@ -54,6 +54,8 @@ class _EditCollectionScreenState extends State<EditCollectionScreen> {
 
     if (widget.collectionDetail.tags != null) {
       tagProvider.saveTags = widget.collectionDetail.tags!;
+    } else {
+      tagProvider.clearTags();
     }
     _userId = widget.collectionDetail.userId;
     _collectionId = widget.collectionDetail.id;
@@ -128,35 +130,37 @@ class _EditCollectionScreenState extends State<EditCollectionScreen> {
       },
     );
     try {
-      if (!(_initialImageName ?? '').contains(_changedImageName ?? '')) {
-        print('바꾸기');
-        print('처음 $_initialImageName');
-        print('바꾼후 $_changedImageName');
+      if (_changedImageName != null) {
+        if (!(_initialImageName ?? '').contains(_changedImageName ?? '')) {
+          print('바꾸기');
+          print('처음 $_initialImageName');
+          print('바꾼후 $_changedImageName');
 
-        if (_initialImageName != null) {
-          print('스토리지 삭제');
-          await ApiService.deleteStorageImages(
-              'collections', [_initialImageName!]);
-        }
+          if (_initialImageName != null) {
+            print('스토리지 삭제');
+            await ApiService.deleteStorageImages(
+                'collections', [_initialImageName!]);
+          }
 
-        if (_changedImageName != 'default') {
-          print('스토리지 등록');
-          await ApiService.copyImageFilePath(
-            'selections',
-            'collections',
-            _changedImageName!,
-            widget.collectionDetail.id,
-          );
+          if (_changedImageName != 'default') {
+            print('스토리지 등록');
+            await ApiService.copyImageFilePath(
+              'selections',
+              'collections',
+              _changedImageName!,
+              widget.collectionDetail.id,
+            );
+            _finalImageFilePath =
+                '${widget.collectionDetail.id}_${_changedImageName}';
+          } else {
+            _finalImageFilePath = null;
+          }
+        } else {
           _finalImageFilePath =
               '${widget.collectionDetail.id}_${_changedImageName}';
-        } else {
-          _finalImageFilePath = null;
         }
-      } else {
-        _finalImageFilePath =
-            '${widget.collectionDetail.id}_${_changedImageName}';
+        print(_finalImageFilePath);
       }
-      print(_finalImageFilePath);
 
       await ApiService.editCollection(
         widget.collectionDetail.id,
